# src/main/java/com/circleci/continuousfooddelivery/api/CartApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (5.3.1).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.circleci.continuousfooddelivery.api;

import com.circleci.continuousfooddelivery.model.Cart;
import com.circleci.continuousfooddelivery.model.Error;
import com.circleci.continuousfooddelivery.model.MenuItem;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;

import javax.servlet.http.HttpServletRequest;
import javax.validation.Valid;
import java.util.Optional;

@javax.annotation.Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2022-01-31T03:17:46.722516300-05:00[America/New_York]")
@Validated
@RestController
@Tag(name = "cart", description = "the cart API")
public interface CartApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /cart : Add a menu item a cart
     * Creates a new item in the cart. Duplicates are allowed
     *
     * @param menuItem Item to add to the cart (required)
     * @return Null response (status code 201)
     * or unexpected error (status code 200)
     */
    @Operation(summary = "Add a menu item into the cart", tags = {"cart",}, responses = {@ApiResponse(responseCode = "201", description = "Null response"), @ApiResponse(responseCode = "200", description = "unexpected error", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class)))})
    @RequestMapping(
            method = RequestMethod.POST,
            value = "/cart",
            produces = {"application/json"},
            consumes = {"application/json"}
    )
    default ResponseEntity<Void> addCartItem(

            @Parameter(name = "Item to add to the cart", required = true) @Valid @RequestBody MenuItem menuItem, HttpServletRequest req) {


        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /cart/{itemId} : Remove item from cart
     * The item must be in the cart. If multiple of same item, call this twice
     *
     * @param itemId The menu item to delete from cart (required)
     * @return Null response (status code 201)
     * or unexpected error (status code 200)
     */
    @Operation(summary = "Remove item from cart", tags = {"cart",}, responses = {@ApiResponse(responseCode = "201", description = "Null response"), @ApiResponse(responseCode = "200", description = "unexpected error", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class)))})
    @RequestMapping(
            method = RequestMethod.DELETE,
            value = "/cart/{itemId}",
            produces = {"application/json"}
    )
    default ResponseEntity<Void> deleteCartItem(@Parameter(name = "itemId", description = "The menu item to delete from cart", required = true)
                                                @PathVariable("itemId") Integer itemId,
                                                HttpServletRequest req) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }


    /**
     * GET /cart : List all cart items
     *
     * @param limit How many items to return at one time (max 100) (optional)
     * @return A paged array of cart items (status code 200)
     * or unexpected error (status code 200)
     */
    @Operation(summary = "List all cart items", tags = {"cart",}, responses = {@ApiResponse(responseCode = "200", description = "A paged array of cart items", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Cart.class))), @ApiResponse(responseCode = "200", description = "unexpected error", content = @Content(mediaType = "application/json", schema = @Schema(implementation = Error.class)))})
    @RequestMapping(
            method = RequestMethod.GET,
            value = "/cart",
            produces = {"application/json"}
    )
    default ResponseEntity<Cart> listCart(@Parameter(name = "limit", description = "How many items to return at one time (max 100)") @Valid @RequestParam(value = "limit", required = false) Integer limit,
                                          HttpServletRequest req
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType : MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "null";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}